# sql, csv: Cache some values between Read() calls to gain performance